- In Revolution surface object, changed the range of default zmin zmax to [0,1]
- Debugged the Lineto in case of the vase form: it used to stop at an arbitray point with unexcepted values.
Th point is that the integration made in the riemannian_turtle_move_forward function by the odeint solver does not manage to converge. I had to decrease rtom and atol from rtol = 1.49012e-8, atol = 1.49012e-8 default values to rtol = 1.49012e-5, atol = 1.49012e-5.

On the web (stack exchange), it is specified that it is a limit of the ODEINT function that don't use events for handling difficult cases.

- Finalized the examples on tropism for pin like structures.